Renaming HGCALReco TICL classes to TICLReco
