+++ DataFormats/HGCalReco/test/BuildFile.xml
@@ -1,4 +1,4 @@
 <bin   name="EtaPhiSearchInTile" file="EtaPhiSearchInTile_t.cpp">
   <use name="catch2"/>
-  <use name="DataFormats/HGCalReco" />
+  <use name="DataFormats/TICLReco" />
 </bin>
