+++ DataFormats/TICLReco/interface/MtdHostCollection.h
@@ -1,8 +1,8 @@
-#ifndef DataFormats_HGCalReco_MtdHostCollection_h
-#define DataFormats_HGCalReco_MtdHostCollection_h
+#ifndef DataFormats_TICLReco_MtdHostCollection_h
+#define DataFormats_TICLReco_MtdHostCollection_h
 
 #include "DataFormats/Portable/interface/PortableHostCollection.h"
-#include "DataFormats/HGCalReco/interface/MtdSoA.h"
+#include "DataFormats/TICLReco/interface/MtdSoA.h"
 
 // MtdSoA in host memory
 using MtdHostCollection = PortableHostCollection<MtdSoA>;
