+++ DataFormats/TICLReco/interface/TICLCandidate.h
@@ -1,9 +1,9 @@
-#ifndef DataFormats_HGCalReco_TICLCandidate_h
-#define DataFormats_HGCalReco_TICLCandidate_h
+#ifndef DataFormats_TICLReco_TICLCandidate_h
+#define DataFormats_TICLReco_TICLCandidate_h
 
 #include "DataFormats/Candidate/interface/LeafCandidate.h"
 #include "DataFormats/Common/interface/Ref.h"
-#include "DataFormats/HGCalReco/interface/Trackster.h"
+#include "DataFormats/TICLReco/interface/Trackster.h"
 #include "DataFormats/Math/interface/Point3D.h"
 #include "DataFormats/TrackReco/interface/Track.h"
 #include "DataFormats/TrackReco/interface/TrackFwd.h"
