+++ DataFormats/TICLReco/interface/TICLLayerTile.h
@@ -1,10 +1,10 @@
 // Authors: Marco Rovere, Felice Pantaleo - marco.rovere@cern.ch, felice.pantaleo@cern.ch
 // Date: 05/2019
 
-#ifndef DataFormats_HGCalReco_TICLLayerTile_h
-#define DataFormats_HGCalReco_TICLLayerTile_h
+#ifndef DataFormats_TICLReco_TICLLayerTile_h
+#define DataFormats_TICLReco_TICLLayerTile_h
 
-#include "DataFormats/HGCalReco/interface/Common.h"
+#include "DataFormats/TICLReco/interface/Common.h"
 #include "DataFormats/Math/interface/normalizedPhi.h"
 
 template <typename T>
