+++ DataFormats/TICLReco/interface/TICLSeedingRegion.h
@@ -2,10 +2,10 @@
 // Emails: felice.pantaleo@cern.ch, marco.rovere@cern.ch
 // Date: 06/2019
 
-#ifndef DataFormats_HGCalReco_TICLSeedingRegion_h
-#define DataFormats_HGCalReco_TICLSeedingRegion_h
+#ifndef DataFormats_TICLSeedingRegion_h
+#define DataFormats_TICLSeedingRegion_h
 
-#include "DataFormats/HGCalReco/interface/Common.h"
+#include "DataFormats/TICLReco/interface/Common.h"
 #include "DataFormats/Math/interface/normalizedPhi.h"
 #include "DataFormats/GeometryVector/interface/GlobalPoint.h"
 #include "DataFormats/GeometryVector/interface/GlobalVector.h"
