+++ DataFormats/TICLReco/src/classes.cc
@@ -1,5 +1,5 @@
 #include "DataFormats/Portable/interface/PortableHostCollectionReadRules.h"
 #include "DataFormats/Portable/interface/PortableHostObjectReadRules.h"
-#include "DataFormats/HGCalReco/interface/MtdHostCollection.h"
+#include "DataFormats/TICLReco/interface/MtdHostCollection.h"
 
 SET_PORTABLEHOSTCOLLECTION_READ_RULES(MtdHostCollection);
